requirements.txt
streamlit
pypdf
python-pptx
scikit-learn
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
chromadb
numpy
sentence-transformers
